reorganized language model lib code
