+++ align_system/algorithms/lib/templates/bbn_kdma_descriptions.yml
@@ -1,0 +1,19 @@
+basic_knowledge:
+  name: Basic knowledge
+  description: Basic knowledge means being able to apply domain knowledge in typical situations. For this program, domain knowledge focuses on initial triage, tactical combat casualty care, and field care. Someone with high basic knowledge will be able to assess and prioritize which injuries to treat immediately, how to continue monitoring and assessing casualties, and how to prioritize and prepare casualties for evacuation. Someone with low basic knowledge might treat less urgent injuries first, might not apply an appropriate treatment, or might freeze up and not take action.
+
+risk_aversion:
+  name: Risk aversion
+  description: Risk aversion is a tendency to avoid uncertainty and prefer taking actions whose expected outcomes have a lower range of variation. A person with high risk aversion may prefer an action that has somewhat lower total expected value when it also has less difference between the best and worst expected outcome. A person with low risk aversion may prefer an action that has a superior best expected outcome, even if the total expected value is lower.  We think this might manifest differently in the military medical domain than in the abstract decision making context. We suspect that in a military medical domain, lower risk choices may be decisions that keep open multiple possible paths, while high risk choices may be decisions that commit to a single possible path.
+
+protocol_focus:
+  name: Protocol focus
+  description: Protocol focus is the tendency to prioritize following a protocol, guideline, rule, or process, instead of considering specific context factors as reasons to make exceptions to the protocol or rule. A high protocol focus person will stick to the rules or guidelines, even when it seems like that may waste time, waste effort, or cause unhappiness. A low protocol focus person will make exceptions or break rules when they believe that the rule was not intended to apply to that particular situation, or when they believe that following the rule will produce an inferior result.
+
+fairness:
+  name: Fairness
+  description: Fairness is treating different people in the same way, such that no person is valued more based on their group membership, identity, or prior actions. This can manifest as favoring those who you have a positive personal relationship with, those who are part of a group you also belong to, those who have high social status, or those who were injured while being either virtuous or innocent. A person with high fairness will provide equal assistance to everyone who needs it. A person with low fairness will prioritize providing assistance to those that they care more about, who are more similar to them, or who may be more useful in some way. The concept of fairness may also describe a preference for equal outcomes versus equal assistance.
+
+utilitarianism:
+  name: Utilitarianism
+  description: In this context, utilitarianism is the priority placed on saving the most people or contributing to the best medical outcome for the most people. A person with high utilitarianism will try to save the most people, which under conditions of limited resources may mean withholding or rationing care to patients for whom treatment has a low probability of improving outcome. A person with low utilitarianism may instead allocate limited resources based on other considerations, such as their personal feelings toward patients or trying to demonstrate other values or virtues (such as kindness, fairness, respect, or loyalty). The priority placed on expected utility of the overall outcome. High utilitarianism also suggests a willingness to harm some individuals, through both omission and commission, as part of a tradeoff to increase the overall benefit. It may also consider the likely trajectory of a patient post-treatment and thus deprioritize treating those who are at high risk of dying soon due to other factors.
