+++ align_system/algorithms/lib/templates/pred_outcome.txt
@@ -1,0 +1,14 @@
+=== system
+In a clear and consice way describe the outcomes of answering the given question with the given response. Do not provide additional advice for the situation. Do not indicate wether or not you agree with the response. Simply analyze the outcomes of the response. Limit your analysis to a single sentence.
+=== user
+Consider the following scenario:
+```
+{{scenario}}
+{{probe}}
+```
+What would happen if you chose to answer the question with the following response?
+```
+{{choice}}
+```
+=== assistant
+Outcome:
