--- align_system/language_model_lib/__init__.py
@@ -1,15 +1,0 @@
-from importlib import reload
-
-def reload_all():
-    # Import the modules inside this function to ensure they're available for reloading
-    
-    from . import util
-    from . import language_model as lm
-    from . import dialog_tokenizer as dt
-    from . import chat_langauge_model as clm
-    from . import llama_2_kdma_predicting_adm as kpa
-
-
-    # Reload in the correct order
-    for module in [util, lm, dt, clm, kpa]:
-        reload(module)
